refactor(hooks): Reorganize and deduplicate hooks
This refactoring improves code organization and reduces duplication:

## High Priority Changes

### 1. Remove Duplicate useDebounce Hook
- Delete duplicate `useDebounce` from apps/web/src/hooks/
- Update 8 files to import from @safe-global/utils/hooks/useDebounce
- Eliminates 30 lines of duplicate code

### 2. Move Feature-Specific Hooks
- Move `useClock` to features/recovery/hooks/ (only used by recovery)
- Move `useIsGeoblockedFeatureEnabled` to features/bridge/hooks/ (only used by bridge)
- Improves code organization and feature cohesion

### 3. Remove Unused Hooks
- Delete `useRemoveAppModal` (no usages found)
- Delete `usePendingActions` (only used in test)
- Removes 74 lines of dead code

## Medium Priority Changes

### 4. Extract Shared Gas Limit Utilities
- Create packages/utils/src/hooks/coreSDK/gasLimitUtils.ts
- Extract shared functions: getEncodedSafeTx, GasMultipliers, incrementByGasMultiplier, getGasLimitForZkSync
- Update both useGasLimit (web) and useDefaultGasLimit (utils) to use shared utilities
- Eliminates 120 lines of duplicated code

### 5. Fix Naming Consistency
- Rename function in useSignerCanPay from useWalletCanPay to useSignerCanPay
- Aligns with filename and interface naming

## Impact Summary
- Files deleted: 4
- Files moved: 2
- Files created: 1 (shared utilities)
- Files updated: 12 (import changes)
- Code reduction: ~224 lines of duplicate/dead code eliminated
- All tests passing (298 suites, 2392 tests)
