+++ apps/web/src/components/common/AddressInput/index.tsx
@@ -18,7 +18,7 @@ import { validatePrefixedAddress } from '@safe-global/utils/utils/validation'
 import { useCurrentChain } from '@/hooks/useChains'
 import useNameResolver from './useNameResolver'
 import { cleanInputValue, parsePrefixedAddress } from '@safe-global/utils/utils/addresses'
-import useDebounce from '@/hooks/useDebounce'
+import useDebounce from '@safe-global/utils/hooks/useDebounce'
 import CaretDownIcon from '@/public/images/common/caret-down.svg'
 import SaveAddressIcon from '@/public/images/common/save-address.svg'
 import classnames from 'classnames'
