+++ apps/web/src/components/common/AddressInput/useNameResolver.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import { useWeb3ReadOnly } from '@/hooks/wallets/web3'
 import useAsync from '@safe-global/utils/hooks/useAsync'
 import { isDomain, resolveName } from '@/services/ens'
-import useDebounce from '@/hooks/useDebounce'
+import useDebounce from '@safe-global/utils/hooks/useDebounce'
 
 const useNameResolver = (
   value?: string,
