+++ apps/web/src/components/tx-flow/flows/UpsertRecovery/RecovererSmartContractWarning.tsx
@@ -7,7 +7,7 @@ import type { ReactElement } from 'react'
 
 import InfoIcon from '@/public/images/notifications/info.svg'
 import { isSmartContractWallet } from '@/utils/wallets'
-import useDebounce from '@/hooks/useDebounce'
+import useDebounce from '@safe-global/utils/hooks/useDebounce'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import { UpsertRecoveryFlowFields } from '.'
 import { sameAddress } from '@safe-global/utils/utils/addresses'
