+++ apps/web/src/hooks/useAddressResolver.ts
@@ -3,7 +3,7 @@ import { useWeb3ReadOnly } from '@/hooks/wallets/web3'
 import { lookupAddress } from '@/services/ens'
 import { useEffect, useMemo } from 'react'
 import useAsync from '@safe-global/utils/hooks/useAsync'
-import useDebounce from './useDebounce'
+import useDebounce from '@safe-global/utils/hooks/useDebounce'
 import { useHasFeature } from './useChains'
 import { FEATURES } from '@safe-global/utils/utils/chains'
 import useChainId from './useChainId'
